Fixes and improvements in answers screens
